+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 * Fixed navigation drawer scrolling
 * Ugraded svelte to finally fix Waypoint detached issue, hence using component Waypoint again
 * Added changelog
+* Add light/dark variants to buttons
+* Refactor buttons to make all classes customizable